net: lib: nrf_cloud: Centralize FOTA support from samples
Move fota support code from the nRF Cloud REST FOTA and
nRF Cloud Multi Service (CoAP mode) samples to an optional
nrf_cloud_fota_poll.c file.

Allow CI to set CONFIG_REST_FOTA_DL_TIMEOUT_MIN,
but use it to set the value of the newer
FOTA_DL_TIMEOUT_MIN.

Rather than managing settings in nrf_cloud_fota.c and
nrf_cloud_fota_poll.c, move them to one place in
nrf_cloud_fota_common.c.

Also fix Coverity defect 201979, and fix broken link
in nrf_cloud_log.h documentation.

Jira: IRIS-6793, IRIS-7159
