+++ samples/cellular/nrf_cloud_multi_service/overlay_coap.conf
@@ -56,6 +56,7 @@ CONFIG_NRF_CLOUD_ALERT=y
 CONFIG_NRF_CLOUD_LOG_DIRECT=y
 CONFIG_NRF_CLOUD_LOG_OUTPUT_LEVEL=3
 CONFIG_NRF_CLOUD_LOG_LOG_LEVEL_INF=y
+CONFIG_NRF_CLOUD_FOTA_POLL=y
 
 # Disable MQTT-specific services; equivalent CoAP versions are used instead.
 CONFIG_NRF_CLOUD_FOTA=n
