+++ samples/cellular/nrf_cloud_rest_fota/prj.conf
@@ -12,6 +12,9 @@ CONFIG_NET_SOCKETS_POSIX_NAMES=y
 CONFIG_NET_NATIVE=n
 CONFIG_NRF_CLOUD_CHECK_CREDENTIALS=y
 
+# FOTA
+CONFIG_NRF_CLOUD_FOTA_POLL=y
+
 # FOTA download
 CONFIG_FOTA_DOWNLOAD=y
 CONFIG_FOTA_DOWNLOAD_PROGRESS_EVT=y
